Refactored quiz table code
