+++ app/javascript/src/translations/en.json
@@ -93,6 +93,9 @@
   "neetoMolecules": {
     "columns": {
       "columns": ""
+    },
+    "common": {
+      "download": "Download report"
     }
   },
   "neetoFilters": {
